fix: apply the conventions review
Config flow: the entry title carried only the model, so two units of the same
product were indistinguishable in the UI; it now includes the address.
Credentials are updated through `async_update_entry` and the existing update
listener rather than `async_update_reload_and_abort`, which schedules a second
reload and is reported as misuse by the pinned core.

The manifest declared `local_push` while every value comes from a GATT read,
and the one user-facing string left in Python is now a translated exception.
The options boundary is read through a cast like `entry.data` already was, and
its declared type matches what the number selector actually writes.

The sensor platform becomes a package, one entity per module, as the style
guide requires once a platform has more than one entity. The product id lives
in `const.py` instead of being repeated per module, and the schema builder no
longer widens its values to `object`.

Tests: the poll timer registration and the picker's already-configured filter
were both deletable with the suite green, and the fixtures carried a real
device's address and a capture of its advertisement into a public repository.

The README, CODE_STYLE and CLAUDE architecture notes all still claimed the
reading was driven by advertisements alone.

+++ custom_components/tuya_ble/coordinator.py
@@ -54,7 +54,7 @@ class TuyaBleDataUpdateCoordinator(
         hass: HomeAssistant,
         config_entry: TuyaBleConfigEntry,
         credentials: TuyaBleCredentials,
-        scan_interval_seconds: int,
+        scan_interval_seconds: float,
     ) -> None:
         """Initialize."""
         self.config_entry = config_entry
